lambda/store_embeddings_lambda/requirements.txt
langchain
boto3
tqdm
faiss-cpu
PyPDF2
python-docx